pyproject.toml
[project]
name = "adforge"
version = "0.1.0"
description = "AI ad forge — playable HTML ads + market-informed creatives, orchestrated with Temporal."
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Orchestration
    "temporalio>=1.7.0",
    # LLM connectors
    "google-genai>=0.3.0",
    "anthropic>=0.40.0",
    "mistralai>=1.2.0",
    # HTTP / data
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "tenacity>=8.5.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    # CLI / UX
    "typer>=0.12.0",
    "rich>=13.7.0",
    # HTML / asset utilities
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "Pillow>=10.3.0",
    "imageio>=2.34.0",
    "imageio-ffmpeg>=0.5.0",
    # Data wrangling
    "pandas>=2.2.0",
]

[project.scripts]
adforge = "adforge.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adforge"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.5.0",
    "pytest>=8.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
